local separation

+++ fabric/build.gradle.kts
@@ -28,7 +28,7 @@ val common: Project = requireNotNull(stonecutter.node.sibling("")) {
 val ci = localProperties.getProperty("ci", "")?.toBoolean() ?: false
 val release = localProperties.getProperty("release", "")?.toBoolean() ?: false
 val nightly = ci && !release
-val buildNumber = mod.build.toInt()
+val buildNumber = mod.build
 
 version = "${mod.version}${if (release) "" else "-dev"}+mc.${minecraft}-${loader}${if (nightly) "-build.${buildNumber}" else ""}"
 group = "${mod.group}.$loader"
